SensorService: Don't always assume magnetometer is present.
The Android 2.3 CDD doesn't mandate that a magnetometer is
present. However, the sensor service always assumes that it is
present, which leads to crashes on devices without one when the
RotationVectorSensor is accessed (such as the Lenovo Ideapad A1).

This commit re-works the sensor service so that it only creates
virtual sensors when the physical sensors that compose them
actually exist.