Update ringer audio focus if ringer volume is changed during ringing.
AudioService only grabs the audio focus for the phone if the ringer
volume is set to something audible. If e.g. ascending ringtone is
enabled and starts at 0, the audio focus wouldn't be grabbed for the
whole ring period, leading to a mixture of played music and the
ringtone. Fix that by grabbing the audio focus if the volume is set away
from 0 while ringing.

Also attenuate music during ringing even when using silent ringtone.
Makes it more obvious to the user that something is going on.